+++ packages/@groupby/elements-sayt-plugin/src/sayt-plugin.ts
@@ -24,7 +24,7 @@ export default class SaytPlugin implements Plugin {
    *
    * @param options The options to instantiate the sayt client with.
    */
-  constructor(options?: SaytOptions) {
+  constructor(options?: SaytPluginOptions) {
     this.sayt = new Sayt(options);
   }
 
@@ -37,4 +37,4 @@ export default class SaytPlugin implements Plugin {
 }
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface SaytOptions extends SaytConfig {}
+export interface SaytPluginOptions extends SaytConfig {}